rename cota settings to cota timeline settings
